+++ IBS/DataAccess/T05Vendor.cs
@@ -13,7 +13,7 @@ public partial class T05Vendor
 
     public string? VendAdd2 { get; set; }
 
-    public byte? VendCityCd { get; set; }
+    public int? VendCityCd { get; set; }
 
     public string? VendContactPer1 { get; set; }
